+++ adapters/adbuttler/adbuttler.go
@@ -1,14 +1,10 @@
 package adbuttler
 
 import (
-	"encoding/json"
 	"fmt"
-	"net/http"
 	"text/template"
 
-	"github.com/mxmCherry/openrtb/v16/openrtb2"
 	"github.com/prebid/prebid-server/adapters"
-	"github.com/prebid/prebid-server/adapters/koddi"
 	"github.com/prebid/prebid-server/config"
 	"github.com/prebid/prebid-server/macros"
 	"github.com/prebid/prebid-server/openrtb_ext"
@@ -21,51 +17,9 @@ type AdButtlerAdapter struct {
 	conversionurl *template.Template
 }
 
-func (a *AdButtlerAdapter) MakeRequests(request *openrtb2.BidRequest, reqInfo *adapters.ExtraRequestInfo) ([]*adapters.RequestData, []error) {
-	host := "localhost"
-	var extension map[string]json.RawMessage
-	var preBidExt openrtb_ext.ExtRequestPrebid
-	var commerceExt koddi.ExtImpCommerce
-	json.Unmarshal(request.Ext, &extension)
-	json.Unmarshal(extension["prebid"], &preBidExt)
-	json.Unmarshal(request.Imp[0].Ext, &commerceExt)
-	endPoint,_ := a.buildEndpointURL(host)
-	errs := make([]error, 0, len(request.Imp))
-
-	reqJSON, err := json.Marshal(request)
-	if err != nil {
-		errs = append(errs, err)
-		return nil, errs
-	}
-
-	headers := http.Header{}
-	headers.Add("Content-Type", "application/json")
-
-	return []*adapters.RequestData{{
-		Method:  "POST",
-		Uri:     endPoint,
-		Body:    reqJSON,
-		Headers: headers,
-	}}, errs
-	
-}
-func (a *AdButtlerAdapter) MakeBids(internalRequest *openrtb2.BidRequest, externalRequest *adapters.RequestData, response *adapters.ResponseData) (*adapters.BidderResponse, []error) {
-	hostName := koddi.GetHostName(internalRequest)
-	if len(hostName) == 0 {
-		hostName = koddi.COMMERCE_DEFAULT_HOSTNAME
-	}
-	iurl, _ := a.buildImpressionURL(hostName) 
-	curl, _ := a.buildClickURL(hostName)
-	purl, _ := a.buildConversionURL(hostName)
-	requestCount := koddi.GetRequestSlotCount(internalRequest)
-	
-	responseF := koddi.GetDummyBids(iurl, curl, purl, "adbuttler", requestCount)
-	return responseF, nil
-}
-
 // Builder builds a new instance of the AdButtler adapter for the given bidder with the given config.
 func Builder(bidderName openrtb_ext.BidderName, config config.Adapter) (adapters.Bidder, error) {
-
+	
 	endpointtemplate, err := template.New("endpointTemplate").Parse(config.Endpoint)
 	if err != nil {
 		return nil, fmt.Errorf("unable to parse endpoint url template: %v", err)
