+++ adapters/adbuttler/adbuttler_request.go
@@ -1,0 +1,57 @@
+package adbuttler
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"strconv"
+	"text/template"
+
+	"github.com/mxmCherry/openrtb/v16/openrtb2"
+	"github.com/prebid/prebid-server/adapters"
+	"github.com/prebid/prebid-server/adapters/koddi"
+	"github.com/prebid/prebid-server/config"
+	"github.com/prebid/prebid-server/macros"
+	"github.com/prebid/prebid-server/openrtb_ext"
+)
+
+
+func (a *AdButtlerAdapter) MakeRequests(request *openrtb2.BidRequest, reqInfo *adapters.ExtraRequestInfo) ([]*adapters.RequestData, []error) {
+	host := "localhost"
+	var extension map[string]json.RawMessage
+	var preBidExt openrtb_ext.ExtRequestPrebid
+	var commerceExt koddi.ExtImpCommerce
+	json.Unmarshal(request.Ext, &extension)
+	json.Unmarshal(extension["prebid"], &preBidExt)
+	json.Unmarshal(request.Imp[0].Ext, &commerceExt)
+	
+	request.TMax = 0
+	customConfig := commerceExt.Bidder.CustomConfig
+	//Nobid := false
+	for _, eachCustomConfig := range customConfig {
+		if *eachCustomConfig.Key == "bidder_timeout"{
+				var timeout int
+				timeout,_ = strconv.Atoi(*eachCustomConfig.Value)
+				request.TMax = int64(timeout)
+		}
+	}
+	endPoint,_ := a.buildEndpointURL(host)
+	errs := make([]error, 0, len(request.Imp))
+
+	reqJSON, err := json.Marshal(request)
+	if err != nil {
+		errs = append(errs, err)
+		return nil, errs
+	}
+
+	headers := http.Header{}
+	headers.Add("Content-Type", "application/json")
+
+	return []*adapters.RequestData{{
+		Method:  "POST",
+		Uri:     endPoint,
+		Body:    reqJSON,
+		Headers: headers,
+	}}, errs
+	
+}
